Fix for loop variables used with lengthof
lengthof returns type size_t, but most for loops used int as a loop
variable.  Fix that.  This avoids possible warnings about
signed/unsigned mismatches under higher warning levels.  (The compiler
will likely optimize these loops beyond recognition, so this shouldn't
affect the generated code much.)

Discussion: https://www.postgresql.org/message-id/flat/d639aede-209f-412b-927a-d38d4848b370%40eisentraut.org

+++ src/backend/access/heap/heapam.c
@@ -8387,7 +8387,7 @@ index_delete_sort(TM_IndexDeleteOp *delstate)
 	StaticAssertDecl(sizeof(TM_IndexDelete) <= 8,
 					 "element size exceeds 8 bytes");
 
-	for (int g = 0; g < lengthof(gaps); g++)
+	for (size_t g = 0; g < lengthof(gaps); g++)
 	{
 		for (int hi = gaps[g], i = hi; i < ndeltids; i++)
 		{
