+++ src/backend/main/main.c
@@ -244,7 +244,7 @@ main(int argc, char *argv[])
 DispatchOption
 parse_dispatch_option(const char *name)
 {
-	for (int i = 0; i < lengthof(DispatchOptionNames); i++)
+	for (size_t i = 0; i < lengthof(DispatchOptionNames); i++)
 	{
 		/*
 		 * Unlike the other dispatch options, "forkchild" takes an argument,
