+++ src/test/modules/test_integerset/test_integerset.c
@@ -116,7 +116,7 @@ test_integerset(PG_FUNCTION_ARGS)
 	test_single_value_and_filler(PG_UINT64_MAX, 1000, 2000);
 
 	/* Test different test patterns, with lots of entries */
-	for (int i = 0; i < lengthof(test_specs); i++)
+	for (size_t i = 0; i < lengthof(test_specs); i++)
 	{
 		test_pattern(&test_specs[i]);
 	}
